requirements.txt
HAP-python
python-miio